bin_tr_height

+++ 9-binary_tree_height.c
@@ -14,8 +14,8 @@ size_t binary_tree_height(const binary_tree_t *tree)
 		return (0);
 	}
 
-	int lh = binary_tree_height(tree->left);
-	int rh = binary_tree_height(tree->right);
+	size_t lh = binary_tree_height(tree->left);
+	size_t rh = binary_tree_height(tree->right);
 
 	if (lh > rh)
 	{
